resume_analyzer/requirements.txt
Flask==2.0.1
Flask-WTF==1.0.1
Werkzeug==2.0.1
google-generative-ai==0.5.4
PyMuPDF==1.23.5
python-dotenv==1.0.0
python-docx==0.8.11